Massive code cleanup and refactor
