--- src/main/kotlin/io/craigmiller160/authmanagementservice/olddto/RoleList.kt
@@ -1,7 +1,0 @@
-package io.craigmiller160.authmanagementservice.olddto
-
-import io.craigmiller160.authmanagementservice.entity.Role
-
-data class RoleList (
-        val roles: List<Role>
-)
